Add embed timeline hook
